SERVER & ClientTest: Création de la méthode epargner

+++ src/main/java/banque/entity/Credit.java
@@ -14,7 +14,7 @@ public class Credit  implements java.io.Serializable {
 
      private Short creditId;
      private CompteEpargne compteEpargne;
-     private Integer montant;
+     private float montant;
      private Short taux;
      private Date dateEmprunt;
      private short duree;
@@ -22,13 +22,13 @@ public class Credit  implements java.io.Serializable {
     public Credit() {
     }
 
-	
+
     public Credit(CompteEpargne compteEpargne, Date dateEmprunt, short duree) {
         this.compteEpargne = compteEpargne;
         this.dateEmprunt = dateEmprunt;
         this.duree = duree;
     }
-    public Credit(CompteEpargne compteEpargne, Integer montant, Short taux, Date dateEmprunt, short duree) {
+    public Credit(CompteEpargne compteEpargne, float montant, Short taux, Date dateEmprunt, short duree) {
        this.compteEpargne = compteEpargne;
        this.montant = montant;
        this.taux = taux;
@@ -40,7 +40,7 @@ public class Credit  implements java.io.Serializable {
     public Short getCreditId() {
         return this.creditId;
     }
-    
+
     public void setCreditId(Short creditId) {
         this.creditId = creditId;
     }
@@ -49,17 +49,17 @@ public class Credit  implements java.io.Serializable {
     public CompteEpargne getCompteEpargne() {
         return this.compteEpargne;
     }
-    
+
     public void setCompteEpargne(CompteEpargne compteEpargne) {
         this.compteEpargne = compteEpargne;
     }
 
     @XmlElement
-    public Integer getMontant() {
+    public float getMontant() {
         return this.montant;
     }
-    
-    public void setMontant(Integer montant) {
+
+    public void setMontant(float montant) {
         this.montant = montant;
     }
 
@@ -67,7 +67,7 @@ public class Credit  implements java.io.Serializable {
     public Short getTaux() {
         return this.taux;
     }
-    
+
     public void setTaux(Short taux) {
         this.taux = taux;
     }
@@ -85,7 +85,7 @@ public class Credit  implements java.io.Serializable {
     public short getDuree() {
         return this.duree;
     }
-    
+
     public void setDuree(short duree) {
         this.duree = duree;
     }
